refactor as per new file structure
